requirements.txt
django-facebook-api>=0.1.1